UnitedStates/src/unitedstates_crawler/__init__.py
"""美国采集器包初始化。"""

from __future__ import annotations

from pathlib import Path

try:
    from dotenv import load_dotenv
except ModuleNotFoundError:  # pragma: no cover - 交付/测试时允许缺少 dotenv
    load_dotenv = None


if load_dotenv is not None:
    load_dotenv(Path(__file__).resolve().parents[2] / ".env")
