UnitedStates/requirements.txt
python-dotenv
curl_cffi
playwright
openai
beautifulsoup4
markdownify
